Refactor options-util
The following points have been taken care of:
- Fixed a linter issue
  jQuery is an unknown global, this has been added to the ignore list
- Renamed isset to isSet
  Method names should use camel casing, the method was renamed so it
  follows this convention

This still leaves two console.log lines which seems really convenient
for debugging purposes but perhaps should be refactored to something
else.

The variable sharedSettingsJSON is not defined, perhaps it is some kind
of global object. If it is, its usage should be documented in the README.